# specs/003-add-this-to/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: DPNR Landing Page API
  version: 1.0.0
  description: API for DPNR course registration, enrollment, and user management
  contact:
    email: Lital2923@gmail.com
  license:
    name: DPNR LTD
    url: https://www.BE-DPNR.com
servers:
  - url: https://api.be-dpnr.com/v1
    description: Production server
  - url: http://localhost:3001/api/v1
    description: Development server
...
paths:
...
  /enrollments:
    post:
      tags: [enrollments]
      summary: Create new enrollment
      description: Submit enrollment application with questionnaire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentRequest'
      responses:
        '201':
          description: Enrollment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '400':
          description: Invalid enrollment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Cohort at capacity or duplicate enrollment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /enrollments/{enrollmentId}:
    get:
      tags: [enrollments]
      summary: Get enrollment details
      description: Retrieve specific enrollment information
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      security:
        - CognitoAuth: []
      responses:
        '200':
          description: Enrollment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'
        '404':
          description: Enrollment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags: [enrollments]
      summary: Update enrollment
      description: Modify enrollment details or status
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnrollmentUpdateRequest'
      security:
        - CognitoAuth: []
      responses:
        '200':
          description: Enrollment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnrollmentResponse'

  # Payment Processing
  /enrollments/{enrollmentId}/payment:
    post:
      tags: [payments]
      summary: Process enrollment payment
      description: Initiate payment through Tranzila gateway
      parameters:
        - name: enrollmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment initiated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Invalid payment data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
  schemas:
...
    CohortResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "December 2025 Cohort"
        startDate:
          type: string
          format: date
          example: "2025-12-01"
        endDate:
          type: string
          format: date
          example: "2026-04-30"
        maxCapacity:
          type: integer
          example: 20
        currentEnrollment:
          type: integer
          example: 15
        status:
          type: string
          enum: [upcoming, open_enrollment, full, in_progress, completed]
        location:
          type: string
          example: "Mazkeret Batya"
        schedule:
          type: string
          example: "Weekly evenings, 1.5-2 hours"
        pricing:
          $ref: '#/components/schemas/CohortPricing'
    CohortPricing:
      type: object
      properties:
        fullPayment:
          type: number
          example: 6400
        fiveInstallments:
          type: number
          example: 1360
        twelveInstallments:
          type: number
          example: 580
        currency:
          type: string
          example: "ILS"
...
    EnrollmentRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - phone
        - cohortId
        - paymentPlan
        - questionnaire
      properties:
        firstName:
          type: string
          minLength: 2
          maxLength: 50
          example: "שרה"
        lastName:
          type: string
          minLength: 2
          maxLength: 50
          example: "כהן"
        email:
          type: string
          format: email
          example: "sarah.cohen@example.com"
        phone:
          type: string
          pattern: '^(\+972|0)5[0-9]-?[0-9]{7}$'
          example: "050-1234567"
        cohortId:
          type: string
          format: uuid
        paymentPlan:
          type: string
          enum: [full, five_installments, twelve_installments]
        preferredLanguage:
          type: string
          enum: [he, en]
          default: he
        questionnaire:
          $ref: '#/components/schemas/QuestionnaireData'
    QuestionnaireData:
      type: object
      required:
        - motivation
        - previousExperience
        - expectations
        - agreedToTerms
        - agreedToPrivacy
      properties:
        motivation:
          type: string
          minLength: 20
          maxLength: 1000
          example: "אני רוצה לפתח את עצמי ולמצוא את הייעוד שלי"
        previousExperience:
          type: boolean
          example: false
        expectations:
          type: string
          minLength: 10
          maxLength: 500
          example: "אני מצפה ללמוד כלים לשינוי והתפתחות אישית"
        referralSource:
          type: string
          maxLength: 200
          example: "חברה המליצה"
        specialNeeds:
          type: string
          maxLength: 500
        agreedToTerms:
          type: boolean
          example: true
        agreedToPrivacy:
          type: boolean
          example: true
        marketingConsent:
          type: boolean
          example: false
    EnrollmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        cohortId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending_payment, active, completed, cancelled, refunded]
        paymentPlan:
          type: string
          enum: [full, five_installments, twelve_installments]
        totalAmount:
          type: number
          example: 6400
        paidAmount:
          type: number
          example: 0
        enrollmentDate:
          type: string
          format: date-time
        questionnaire:
          $ref: '#/components/schemas/QuestionnaireData'
        cohort:
          $ref: '#/components/schemas/CohortResponse'
    EnrollmentUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [cancelled]
        questionnaire:
          $ref: '#/components/schemas/QuestionnaireData'
    PaymentRequest:
      type: object
      required:
        - amount
        - installmentNumber
      properties:
        amount:
          type: number
          minimum: 1
          example: 6400
        installmentNumber:
          type: integer
          minimum: 1
          maximum: 12
          example: 1
        returnUrl:
          type: string
          format: uri
          example: "https://www.be-dpnr.com/payment/success"
        cancelUrl:
          type: string
          format: uri
          example: "https://www.be-dpnr.com/payment/cancel"
    PaymentResponse:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        tranzilaUrl:
          type: string
          format: uri
          description: Redirect URL for Tranzila payment page
        amount:
          type: number
          example: 6400
        currency:
          type: string
          example: "ILS"
        expiresAt:
          type: string
          format: date-time
          description: Payment session expiration
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Required field missing"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: "/api/v1/enrollments"